+++ test/src/Release.test.js
@@ -1,17 +1,17 @@
-import ReleaseDeployer from "../../src/release/ReleaseDeployer"
+import Release from '../../src/release/Release'
 
 require('chai')
   .use(require('chai-as-promised'))
   .use(require('chai-bignumber')(web3.BigNumber))
   .should()
 
-contract('ReleaseWrapper', ([_, owner]) => {
+contract('Release', ([_, owner]) => {
   const txParams = { from: owner }
   const contracts = [{ alias: 'DummyImplementation', name: 'DummyImplementation' }]
 
-  describe('call', function () {
+  describe('deploy', function () {
     beforeEach(async function () {
-      this.release = await ReleaseDeployer.call(contracts, txParams)
+      this.release = await Release.deploy(contracts, txParams)
     })
 
     it('has an owner', async function () {
@@ -25,6 +25,15 @@ contract('ReleaseWrapper', ([_, owner]) => {
     })
 
     it('can tell the implementation of a contract', async function () {
+      (await this.release.getImplementation('DummyImplementation')).should.not.be.zero
+    })
+
+    it('deploys a new release', async function () {
+      this.release.address().should.not.be.null
+      this.release.owner().should.eventually.be.equal(owner)
+    })
+
+    it('includes the given contracts', async function () {
       (await this.release.getImplementation('DummyImplementation')).should.not.be.zero
     })
   })
